backend/pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "docx>=0.2.4",
    "fastapi>=0.121.2",
    "google-genai>=1.51.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.3.0",
    "python-docx>=1.2.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

include = ["."]
exclude = [
    ".venv",
    "venv",
]

